requirements.txt
fastapi==0.109.2
uvicorn==0.15.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
pymongo>=4.6.1
httpx==0.24.1
streamlit==1.24.0
sentence-transformers>=2.2.2
transformers>=4.34.0
torch>=2.0.0
huggingface-hub>=0.19.0
pinecone-client==3.0.0
python-dotenv==0.19.2
beautifulsoup4==4.11.1
PyPDF2==3.0.1
python-docx==0.8.11
pydantic>=2.0.3
email-validator==2.0.0.post2 